fixed bug

+++ packages/contracts/contracts/Interfaces/ISubscription.sol
@@ -3,6 +3,6 @@
 pragma solidity 0.6.11;
 
 abstract contract ISubscription {
-    function subscribe(uint128 _minRatio) public virtual;
+    function subscribe(uint _minRatio) public virtual;
     function unsubscribe() public virtual;
 }
